requirements.txt
# Streamlit Frontend
streamlit

# Azure OpenAI & LangChain
langchain-openai
openai

# Vector Store (for future RAG chat feature)
chromadb
langchain-chroma
rank-bm25

# Environment & Utils
python-dotenv
requests
